refactor: 사이드바 리팩토링
